Mellanox common code restructuring. MSN2100, MSN2410, MSN2700 platforms support.
