+++ packages/platforms/mellanox/any/src/mlnx_common/.module
@@ -1,0 +1,1 @@
+name: mlnx_common
