+++ packages/platforms/mellanox/any/src/mlnx_common/Makefile
@@ -1,0 +1,9 @@
+###############################################################################
+#
+#
+#
+###############################################################################
+include $(ONL)/make/config.mk
+MODULE := mlnx_common
+AUTOMODULE := mlnx_common
+include $(BUILDER)/definemodule.mk
