+++ packages/platforms/mellanox/any/src/mlnx_common/README
@@ -1,0 +1,5 @@
+###############################################################################
+#
+# mlnx_common README
+#
+###############################################################################
