+++ packages/platforms/mellanox/any/src/mlnx_common/module/auto/make.mk
@@ -1,0 +1,8 @@
+###############################################################################
+#
+# mlnx_common Autogeneration
+#
+###############################################################################
+mlnx_common_AUTO_DEFS := module/auto/mlnx_common.yml
+mlnx_common_AUTO_DIRS := module/inc/mlnx_common module/src
+include $(BUILDER)/auto.mk
