+++ packages/platforms/mellanox/any/src/mlnx_common/module/inc/mlnx_common/mlnx_common.x
@@ -1,0 +1,12 @@
+/**************************************************************************//**
+ *
+ *
+ *
+ *****************************************************************************/
+#include <mlnx_common/mlnx_common_config.h>
+
+/* <--auto.start.xmacro(ALL).define> */
+/* <auto.end.xmacro(ALL).define> */
+
+/* <--auto.start.xenum(ALL).define> */
+/* <auto.end.xenum(ALL).define> */
