+++ packages/platforms/mellanox/any/src/mlnx_common/module/inc/mlnx_common/mlnx_common_dox.h
@@ -1,0 +1,26 @@
+/**************************************************************************//**
+ *
+ * mlnx_common Doxygen Header
+ *
+ *****************************************************************************/
+#ifndef __MLNX_COMMON_DOX_H__
+#define __MLNX_COMMON_DOX_H__
+
+/**
+ * @defgroup mlnx_common mlnx_common - mlnx_common Description
+ *
+
+The documentation overview for this module should go here.
+
+ *
+ * @{
+ *
+ * @defgroup mlnx_common-mlnx_common Public Interface
+ * @defgroup mlnx_common-config Compile Time Configuration
+ * @defgroup mlnx_common-porting Porting Macros
+ *
+ * @}
+ *
+ */
+
+#endif /* __MLNX_COMMON_DOX_H__ */
