+++ packages/platforms/mellanox/any/src/mlnx_common/module/src/Makefile
@@ -1,0 +1,8 @@
+###############################################################################
+#
+# Local source generation targets.
+#
+###############################################################################
+
+ucli:
+	@../../../../tools/uclihandlers.py mlnx_common_ucli.c
