+++ packages/platforms/mellanox/any/src/mlnx_common/module/src/make.mk
@@ -1,0 +1,9 @@
+###############################################################################
+#
+#
+#
+###############################################################################
+
+LIBRARY := mlnx_common
+$(LIBRARY)_SUBDIR := $(dir $(lastword $(MAKEFILE_LIST)))
+include $(BUILDER)/lib.mk
