+++ packages/platforms/mellanox/any/src/mlnx_common/module/src/mlnx_common_enums.c
@@ -1,0 +1,9 @@
+/**************************************************************************//**
+ *
+ *
+ *
+ *****************************************************************************/
+#include <mlnx_common/mlnx_common_config.h>
+
+/* <--auto.start.enum(ALL).source> */
+/* <auto.end.enum(ALL).source> */
