+++ packages/platforms/mellanox/any/src/mlnx_common/module/src/mlnx_common_log.c
@@ -1,0 +1,17 @@
+/**************************************************************************//**
+ *
+ *
+ *
+ *****************************************************************************/
+#include <mlnx_common/mlnx_common_config.h>
+
+#include "mlnx_common_log.h"
+/*
+ * mlnx_common log struct.
+ */
+AIM_LOG_STRUCT_DEFINE(
+                      MLNX_COMMON_CONFIG_LOG_OPTIONS_DEFAULT,
+                      MLNX_COMMON_CONFIG_LOG_BITS_DEFAULT,
+                      NULL, /* Custom log map */
+                      MLNX_COMMON_CONFIG_LOG_CUSTOM_BITS_DEFAULT
+                     );
