+++ packages/platforms/mellanox/any/src/mlnx_common/module/src/mlnx_common_log.h
@@ -1,0 +1,12 @@
+/**************************************************************************//**
+ *
+ *
+ *
+ *****************************************************************************/
+#ifndef __MLNX_COMMON_LOG_H__
+#define __MLNX_COMMON_LOG_H__
+
+#define AIM_LOG_MODULE_NAME mlnx_common
+#include <AIM/aim_log.h>
+
+#endif /* __MLNX_COMMON_LOG_H__ */
