+++ packages/platforms/mellanox/any/src/mlnx_common/module/src/mlnx_common_psui.c
@@ -19,7 +19,7 @@
  * </bsn.cl>
  ************************************************************
  *
- *
+ * PSU Platform Implementation Defaults.
  *
  ***********************************************************/
 #include <stdio.h>
@@ -27,13 +27,16 @@
 #include <onlplib/file.h>
 #include <onlplib/mmap.h>
 #include <onlp/platformi/psui.h>
-#include "platform_lib.h"
+#include <mlnx_common/mlnx_common.h>
+#include "mlnx_common_log.h"
 
 #define PSU_STATUS_PRESENT   1
 #define PSU_CABLE_PRESENT    1
 
 #define PSU_NODE_MAX_INT_LEN  8
 #define PSU_NODE_MAX_PATH_LEN 64
+
+#define PSU_MODEL	"POW000167"
 
 #define VALIDATE(_id)                           \
     do {                                        \
@@ -42,35 +45,31 @@
         }                                       \
     } while(0)
 
-static int
+int
 psu_module_info_get(int id, char *node, int *value)
 {
-    int len, ret = 0;
-    char buf[PSU_NODE_MAX_INT_LEN + 1] = {0};
+    int ret = 0;
 
     *value = 0;
 
-    ret = onlp_file_read((uint8_t*)buf, sizeof(buf), &len,
-    		PSU_MODULE_PREFIX, id, node);
-    if (ret == 0) {
-        *value = atoi(buf);
+    ret = onlp_file_read_int(value, PSU_MODULE_PREFIX, id, node);
+    if (ret < 0) {
+        return ONLP_STATUS_E_INTERNAL;
     }
 
     return ret;
 }
 
-static int
+int
 psu_power_info_get(int id, char *node, int *value)
 {
-    int len, ret = 0;
-    char buf[PSU_NODE_MAX_INT_LEN + 1] = {0};
+    int ret = 0;
 
     *value = 0;
 
-    ret = onlp_file_read((uint8_t*)buf, sizeof(buf), &len,
-    		PSU_POWER_PREFIX, id, node);
-    if (ret == 0) {
-        *value = atoi(buf);
+    ret = onlp_file_read_int(value, PSU_POWER_PREFIX, id, node);
+    if (ret < 0) {
+        return ONLP_STATUS_E_INTERNAL;
     }
 
     return ret;
@@ -82,12 +81,63 @@ onlp_psui_init(void)
     return ONLP_STATUS_OK;
 }
 
+/*
+ * Get all information about the given PSU oid.
+ */
+static onlp_psu_info_t pinfo[] =
+{
+    { }, /* Not used */
+    {
+        { ONLP_PSU_ID_CREATE(PSU1_ID), "PSU-1", 0 },
+    },
+    {
+        { ONLP_PSU_ID_CREATE(PSU2_ID), "PSU-2", 0 },
+    }
+};
+
 static int
-_psu_info_get(onlp_psu_info_t* info)
+_psu_info_get_type1(onlp_psu_info_t* info)
+{
+    int val   = 0;
+    int index = ONLP_OID_ID_GET(info->hdr.id);
+    const char psu_model[]=PSU_MODEL;
+
+    strncpy(info->model, psu_model, sizeof(info->model));
+
+    /* Set capability */
+    info->caps = ONLP_PSU_CAPS_AC;
+
+    if (info->status & ONLP_PSU_STATUS_FAILED) {
+        return ONLP_STATUS_OK;
+    }
+
+    /* Read voltage, current and power */
+    if (psu_power_info_get(index, "volt", &val) == 0) {
+        info->mvout = val;
+        info->caps |= ONLP_PSU_CAPS_VOUT;
+    }
+
+    if (psu_power_info_get(index, "curr", &val) == 0) {
+        info->miout = val;
+        info->caps |= ONLP_PSU_CAPS_IOUT;
+    }
+
+    info->mpout = info->mvout * info->miout;
+    info->caps |= ONLP_PSU_CAPS_POUT;
+
+    info->mpin = ((int)(info->mpout / 91)) * 100;
+    info->caps |= ONLP_PSU_CAPS_PIN;
+
+    return ONLP_STATUS_OK;
+}
+
+static int
+_psu_info_get_type2(onlp_psu_info_t* info)
 {
     int val   = 0;
     int index = ONLP_OID_ID_GET(info->hdr.id);
 
+    mlnx_platform_info_t* mlnx_platform_info = get_platform_info();
     /* Set capability
      */
     info->caps = ONLP_PSU_CAPS_AC;
@@ -97,12 +147,12 @@ _psu_info_get(onlp_psu_info_t* info)
     }
 
     /* Set the associated oid_table */
-    info->hdr.coids[0] = ONLP_FAN_ID_CREATE(index + CHASSIS_FAN_COUNT);
-    info->hdr.coids[1] = ONLP_THERMAL_ID_CREATE(index + CHASSIS_THERMAL_COUNT);
+    info->hdr.coids[0] = ONLP_FAN_ID_CREATE(index + mlnx_platform_info->fan_num);
+    info->hdr.coids[1] = ONLP_THERMAL_ID_CREATE(index + mlnx_platform_info->thermal_num);
 
     /* Read voltage, current and power */
     if (psu_power_info_get(index, "volt_in", &val) == 0 &&
-        0 != val) {
+            0 != val) {
         info->mvin = val;
         info->caps |= ONLP_PSU_CAPS_VIN;
 
@@ -138,19 +188,20 @@ _psu_info_get(onlp_psu_info_t* info)
     return psu_read_eeprom(index, info, NULL);
 }
 
-/*
- * Get all information about the given PSU oid.
- */
-static onlp_psu_info_t pinfo[] =
+int _psu_info_get(onlp_psu_info_t* info)
 {
-    { }, /* Not used */
-    {
-        { ONLP_PSU_ID_CREATE(PSU1_ID), "PSU-1", 0 },
-    },
-    {
-        { ONLP_PSU_ID_CREATE(PSU2_ID), "PSU-2", 0 },
+    int res;
+    mlnx_platform_info_t* mlnx_platform_info = get_platform_info();
+    switch(mlnx_platform_info->psu_type) {
+      case PSU_TYPE_1:
+	  res=_psu_info_get_type1(info);
+	  break;
+      case PSU_TYPE_2:
+	  res=_psu_info_get_type2(info);
+	  break;
     }
-};
+    return res;
+}
 
 int
 onlp_psui_info_get(onlp_oid_t id, onlp_psu_info_t* info)
@@ -164,37 +215,39 @@ onlp_psui_info_get(onlp_oid_t id, onlp_psu_info_t* info)
     memset(info, 0, sizeof(onlp_psu_info_t));
     *info = pinfo[index]; /* Set the onlp_oid_hdr_t */
 
-    /* Get the present state */
-    if (psu_module_info_get(index, "status", &val) != 0) {
-    	AIM_LOG_ERROR("Unable to read PSU(%d) node(psu_present)\r\n", index);
-    }
+    mlnx_platform_info_t* mlnx_platform_info = get_platform_info();
+       if(mlnx_platform_info->psu_fixed == false) {
+	  /* Get the present state */
+	  if (psu_module_info_get(index, "status", &val) != 0) {
+	      AIM_LOG_ERROR("Unable to read PSU(%d) node(psu_present)\r\n", index);
+	  }
 
-    if (val != PSU_STATUS_PRESENT) {
-        info->status &= ~ONLP_PSU_STATUS_PRESENT;
-        info->status |= ONLP_PSU_STATUS_UNPLUGGED;
-        return ONLP_STATUS_OK;
-    }
-    else
-    	info->status |= ONLP_PSU_STATUS_PRESENT;
+	  if (val != PSU_STATUS_PRESENT) {
+	      info->status &= ~ONLP_PSU_STATUS_PRESENT;
+	      info->status |= ONLP_PSU_STATUS_UNPLUGGED;
+	      return ONLP_STATUS_OK;
+	  }
+	  else
+	      info->status |= ONLP_PSU_STATUS_PRESENT;
+	 }
+       else {
+	 /* Fixed system, PSU is always present */
+	     info->status |= ONLP_PSU_STATUS_PRESENT;
+       }
 
     /* Get the cable preset state */
     if (psu_module_info_get(index, "pwr_status", &val) != 0) {
-    	AIM_LOG_ERROR("Unable to read PSU(%d) node(cable_present)\r\n", index);
+        AIM_LOG_ERROR("Unable to read PSU(%d) node(cable_present)\r\n", index);
     }
 
     if (val != PSU_CABLE_PRESENT) {
         info->status |= ONLP_PSU_STATUS_UNPLUGGED;
         return ONLP_STATUS_OK;
     }
-
+    if(mlnx_platform_info->psu_fixed == false) {
+	info->status |= ONLP_PSU_STATUS_PRESENT;
+    }
     ret = _psu_info_get(info);
 
     return ret;
 }
-
-int
-onlp_psui_ioctl(onlp_oid_t pid, va_list vargs)
-{
-    return ONLP_STATUS_E_UNSUPPORTED;
-}
-
