+++ packages/platforms/mellanox/any/src/mlnx_common/utest/_make.mk
@@ -1,0 +1,8 @@
+###############################################################################
+#
+# mlnx_common Unit Test Makefile.
+#
+###############################################################################
+UMODULE := mlnx_common
+UMODULE_SUBDIR := $(dir $(lastword $(MAKEFILE_LIST)))
+include $(BUILDER)/utest.mk
