+++ packages/platforms/mellanox/x86-64/x86-64-mlnx-msn2100/onlp/builds/lib/Makefile
@@ -26,7 +26,7 @@ include $(ONL)/make/config.amd64.mk
 MODULE := libonlp-x86-64-mlnx-msn2100
 include $(BUILDER)/standardinit.mk
 
-DEPENDMODULES := AIM IOF x86_64_mlnx_msn2100 onlplib
+DEPENDMODULES := AIM IOF mlnx_common x86_64_mlnx_msn2100 onlplib
 DEPENDMODULE_HEADERS := sff
 
 include $(BUILDER)/dependmodules.mk
@@ -37,6 +37,7 @@ include $(BUILDER)/so.mk
 .DEFAULT_GOAL := $(SHAREDLIB)
 
 GLOBAL_CFLAGS += -I$(onlp_BASEDIR)/module/inc
+GLOBAL_CFLAGS += -I$(mlnx_common_BASEDIR)/module/inc
 GLOBAL_CFLAGS += -DAIM_CONFIG_INCLUDE_MODULES_INIT=1
 GLOBAL_CFLAGS += -fPIC
 GLOBAL_LINK_LIBS += -lpthread
