+++ packages/platforms/mellanox/x86-64/x86-64-mlnx-msn2700/onlp/builds/src/module/src/ledi.c
@@ -30,101 +30,9 @@
 #include <onlplib/file.h>
 #include <onlp/platformi/ledi.h>
 #include "platform_lib.h"
+#include <mlnx_common/mlnx_common.h>
 
-#define prefix_path "/bsp/led/led_"
-#define driver_value_len 50
-
-#define LED_MODE_OFF         "none"
-#define LED_MODE_GREEN       "green"
-#define LED_MODE_RED         "red"
-#define LED_MODE_BLUE        "blue"
-#define LED_MODE_GREEN_BLINK "green_blink"
-#define LED_MODE_RED_BLINK   "red_blink"
-#define LED_MODE_BLUE_BLINK  "blue_blink"
-#define LED_MODE_AUTO        "cpld_control"
-
-#define LED_BLINK_PERIOD	"100"
-#define LED_ON			"1"
-#define LED_OFF			"0"
-#define LED_BLINK_PERIOD_LEN	3
-#define LED_MODE_LEN		1
-
-#define VALIDATE(_id)                           \
-    do {                                        \
-        if(!ONLP_OID_IS_LED(_id)) {             \
-            return ONLP_STATUS_E_INVALID;       \
-        }                                       \
-    } while(0)
-
-/* LED related data
- */
-
-typedef struct led_light_mode_map {
-    enum onlp_led_id id;
-    char* driver_led_mode;
-    enum onlp_led_mode_e onlp_led_mode;
-} led_light_mode_map_t;
-
-led_light_mode_map_t led_map[] = {
-{LED_SYSTEM, LED_MODE_OFF,         ONLP_LED_MODE_OFF},
-{LED_SYSTEM, LED_MODE_GREEN,       ONLP_LED_MODE_GREEN},
-{LED_SYSTEM, LED_MODE_RED,         ONLP_LED_MODE_RED},
-{LED_SYSTEM, LED_MODE_RED_BLINK,   ONLP_LED_MODE_RED_BLINKING},
-{LED_SYSTEM, LED_MODE_GREEN_BLINK, ONLP_LED_MODE_GREEN_BLINKING},
-{LED_SYSTEM, LED_MODE_AUTO,        ONLP_LED_MODE_AUTO},
-
-{LED_FAN1, LED_MODE_OFF,         ONLP_LED_MODE_OFF},
-{LED_FAN1, LED_MODE_GREEN,       ONLP_LED_MODE_GREEN},
-{LED_FAN1, LED_MODE_RED,         ONLP_LED_MODE_RED},
-{LED_FAN1, LED_MODE_RED_BLINK,   ONLP_LED_MODE_RED_BLINKING},
-{LED_FAN1, LED_MODE_GREEN_BLINK, ONLP_LED_MODE_GREEN_BLINKING},
-{LED_FAN1, LED_MODE_AUTO,        ONLP_LED_MODE_AUTO},
-
-{LED_FAN2, LED_MODE_OFF,         ONLP_LED_MODE_OFF},
-{LED_FAN2, LED_MODE_GREEN,       ONLP_LED_MODE_GREEN},
-{LED_FAN2, LED_MODE_RED,         ONLP_LED_MODE_RED},
-{LED_FAN2, LED_MODE_RED_BLINK,   ONLP_LED_MODE_RED_BLINKING},
-{LED_FAN2, LED_MODE_GREEN_BLINK, ONLP_LED_MODE_GREEN_BLINKING},
-{LED_FAN2, LED_MODE_AUTO,        ONLP_LED_MODE_AUTO},
-
-{LED_FAN3, LED_MODE_OFF,         ONLP_LED_MODE_OFF},
-{LED_FAN3, LED_MODE_GREEN,       ONLP_LED_MODE_GREEN},
-{LED_FAN3, LED_MODE_RED,         ONLP_LED_MODE_RED},
-{LED_FAN3, LED_MODE_RED_BLINK,   ONLP_LED_MODE_RED_BLINKING},
-{LED_FAN3, LED_MODE_GREEN_BLINK, ONLP_LED_MODE_GREEN_BLINKING},
-{LED_FAN3, LED_MODE_AUTO,        ONLP_LED_MODE_AUTO},
-
-{LED_FAN4, LED_MODE_OFF,         ONLP_LED_MODE_OFF},
-{LED_FAN4, LED_MODE_GREEN,       ONLP_LED_MODE_GREEN},
-{LED_FAN4, LED_MODE_RED,         ONLP_LED_MODE_RED},
-{LED_FAN4, LED_MODE_RED_BLINK,   ONLP_LED_MODE_RED_BLINKING},
-{LED_FAN4, LED_MODE_GREEN_BLINK, ONLP_LED_MODE_GREEN_BLINKING},
-{LED_FAN4, LED_MODE_AUTO,        ONLP_LED_MODE_AUTO},
-
-{LED_PSU, LED_MODE_OFF,         ONLP_LED_MODE_OFF},
-{LED_PSU, LED_MODE_GREEN,       ONLP_LED_MODE_GREEN},
-{LED_PSU, LED_MODE_RED,         ONLP_LED_MODE_RED},
-{LED_PSU, LED_MODE_RED_BLINK,   ONLP_LED_MODE_RED_BLINKING},
-{LED_PSU, LED_MODE_GREEN_BLINK, ONLP_LED_MODE_GREEN_BLINKING},
-{LED_PSU, LED_MODE_AUTO,        ONLP_LED_MODE_AUTO}
-};
-
-typedef struct led_colors {
-    enum onlp_led_id id;
-    const char* color1;
-    const char* color2;
-} led_colors_t;
-
-static led_colors_t led_colors_map[] = {
-    {LED_SYSTEM, "green", "red"},
-    {LED_FAN1, "green", "red"},
-    {LED_FAN2, "green", "red"},
-    {LED_FAN3, "green", "red"},
-    {LED_FAN4, "green", "red"},
-    {LED_PSU, "green", "red"},
-};
-
-static char file_names[][10] =  /* must map with onlp_led_id */
+static char* file_names[] =  /* must map with onlp_led_id */
 {
     "reserved",
     "status",
@@ -179,213 +87,11 @@ static onlp_led_info_t linfo[] =
     }
 };
 
-static int driver_to_onlp_led_mode(enum onlp_led_id id, char* driver_led_mode)
-{
-    char *pos;
-    int i, nsize = sizeof(led_map)/sizeof(led_map[0]);
-
-    if ((pos=strchr(driver_led_mode, '\n')) != NULL)
-        *pos = '\0';
-    for (i = 0; i < nsize; i++)
-    {
-        if (id == led_map[i].id &&
-            !strncmp(led_map[i].driver_led_mode, driver_led_mode, driver_value_len))
-        {
-            return led_map[i].onlp_led_mode;
-        }
-    }
-
-    return 0;
-}
-
-static char* onlp_to_driver_led_mode(enum onlp_led_id id, onlp_led_mode_t onlp_led_mode)
-{
-    int i, nsize = sizeof(led_map)/sizeof(led_map[0]);
-
-    for (i = 0; i < nsize; i++)
-    {
-        if (id == led_map[i].id && onlp_led_mode == led_map[i].onlp_led_mode)
-        {
-            return led_map[i].driver_led_mode;
-        }
-    }
-
-    return LED_MODE_OFF;
-}
-
-static int led_set_mode(onlp_oid_t id, onlp_led_mode_t mode)
-{
-    int  local_id = ONLP_OID_ID_GET(id);
-    char color[10]={0};
-    int blinking = 0;
-
-    switch (mode) {
-    case ONLP_LED_MODE_RED_BLINKING:
-        strcpy(color, "red");
-        blinking = 1;
-        break;
-    case ONLP_LED_MODE_GREEN_BLINKING:
-        strcpy(color, "green");
-        blinking = 1;
-        break;
-    case ONLP_LED_MODE_BLUE_BLINKING:
-        strcpy(color, "blue");
-        blinking = 1;
-        break;
-    case ONLP_LED_MODE_YELLOW_BLINKING:
-        strcpy(color, "yellow");
-        blinking = 1;
-        break;
-    case ONLP_LED_MODE_RED:
-    	strcpy(color, "red");
-    	break;
-    case ONLP_LED_MODE_GREEN:
-    	strcpy(color, "green");
-    	break;
-    case ONLP_LED_MODE_BLUE:
-    	strcpy(color, "blue");
-    	break;
-    case ONLP_LED_MODE_YELLOW:
-    	strcpy(color, "yellow");
-        break;
-    default:
-        return ONLP_STATUS_E_PARAM;
-    }
-
-    if (blinking) {
-    	onlp_file_write((uint8_t*)LED_BLINK_PERIOD, LED_BLINK_PERIOD_LEN,
-    			"%s%s_%s_delay_off", prefix_path, file_names[local_id], color);
-    	onlp_file_write((uint8_t*)LED_BLINK_PERIOD, LED_BLINK_PERIOD_LEN,
-    	    	"%s%s_%s_delay_on", prefix_path, file_names[local_id], color);
-    }
-    onlp_file_write((uint8_t*)LED_ON, LED_MODE_LEN,
-    		"%s%s_%s", prefix_path, file_names[local_id], color);
-
-    return ONLP_STATUS_OK;
-}
-
-/*
- * This function will be called prior to any other onlp_ledi_* functions.
- */
 int
 onlp_ledi_init(void)
 {
-    /*
-     * ONLPD calls it too early before all BSP insfrastructure is set
-     */
-
+    mlnx_platform_info_t* mlnx_platform_info = get_platform_info();
+    mlnx_platform_info->linfo = linfo;
+    mlnx_platform_info->led_fnames = file_names;
     return ONLP_STATUS_OK;
 }
-
-int
-onlp_ledi_info_get(onlp_oid_t id, onlp_led_info_t* info)
-{
-    int  len, local_id = 0;
-    uint8_t data[driver_value_len] = {0};
-
-    VALIDATE(id);
-
-    local_id = ONLP_OID_ID_GET(id);
-
-    /* Set the onlp_oid_hdr_t and capabilities */
-    *info = linfo[ONLP_OID_ID_GET(id)];
-
-    /* Get LED mode */
-    if (onlp_get_kernel_ver() >= KERNEL_VERSION(4,9,30)) {
-        char* cmd = aim_fstrdup("%s%s_state", prefix_path, file_names[local_id]);
-        if(system(cmd) != 0) {
-            aim_free(cmd);
-            return ONLP_STATUS_E_INTERNAL;
-        }
-        aim_free(cmd);
-    }
-
-    if (onlp_file_read(data, sizeof(data), &len, "%s%s",
-    		prefix_path, file_names[local_id]) != 0) {
-        return ONLP_STATUS_E_INTERNAL;
-    }
-
-    info->mode = driver_to_onlp_led_mode(local_id, (char*)data);
-
-    /* Set the on/off status */
-    if (info->mode != ONLP_LED_MODE_OFF) {
-        info->status |= ONLP_LED_STATUS_ON;
-    }
-
-    return ONLP_STATUS_OK;
-}
-
-/*
- * Turn an LED on or off.
- *
- * This function will only be called if the LED OID supports the ONOFF
- * capability.
- *
- * What 'on' means in terms of colors or modes for multimode LEDs is
- * up to the platform to decide. This is intended as baseline toggle mechanism.
- */
-int
-onlp_ledi_set(onlp_oid_t id, int on_or_off)
-{
-    VALIDATE(id);
-
-    if (!on_or_off) {
-    	if (onlp_get_kernel_ver() < KERNEL_VERSION(4,9,30))
-        return onlp_ledi_mode_set(id, ONLP_LED_MODE_OFF);
-    	else {
-    	    int i, nsize = sizeof(led_colors_map)/sizeof(led_colors_map[0]);
-    	    for (i = 0; i < nsize; i++)
-    	    {
-    	        if (id == led_colors_map[i].id)
-    	            break;
-    	    }
-    	    if (led_colors_map[i].color1)
-    		onlp_file_write((uint8_t*)LED_OFF, LED_MODE_LEN,
-    		    		"%s%s_%s", prefix_path, file_names[id], led_colors_map[i].color1);
-    	}
-    }
-
-    return ONLP_STATUS_E_UNSUPPORTED;
-}
-
-/*
- * This function puts the LED into the given mode. It is a more functional
- * interface for multimode LEDs.
- *
- * Only modes reported in the LED's capabilities will be attempted.
- */
-int
-onlp_ledi_mode_set(onlp_oid_t id, onlp_led_mode_t mode)
-{
-    int  local_id;
-    char* driver_led_mode;
-    int nbytes;
-
-    VALIDATE(id);
-
-    if (onlp_get_kernel_ver() < KERNEL_VERSION(4,9,30)) {
-    local_id = ONLP_OID_ID_GET(id);
-    	driver_led_mode = onlp_to_driver_led_mode(local_id, mode);
-    	nbytes = strnlen(driver_led_mode, driver_value_len);
-        if (onlp_file_write((uint8_t*)driver_led_mode, nbytes,
-        		"%s%s", prefix_path, file_names[local_id]) != 0) {
-        return ONLP_STATUS_E_INTERNAL;
-        }
-    } else {
-    	if (led_set_mode(id, mode) != 0) {
-    	    return ONLP_STATUS_E_INTERNAL;
-    	}
-    }
-
-    return ONLP_STATUS_OK;
-}
-
-/*
- * Generic LED ioctl interface.
- */
-int
-onlp_ledi_ioctl(onlp_oid_t id, va_list vargs)
-{
-    return ONLP_STATUS_E_UNSUPPORTED;
-}
-
