+++ packages/platforms/mellanox/x86-64/x86-64-mlnx-msn2700/onlp/builds/src/module/src/platform_lib.h
@@ -25,8 +25,6 @@
 #ifndef __PLATFORM_LIB_H__
 #define __PLATFORM_LIB_H__
 
-#include <onlp/fan.h>
-#include <onlp/psu.h>
 #include "x86_64_mlnx_msn2700_log.h"
 
 #define CHASSIS_PSU_COUNT           2
@@ -34,43 +32,8 @@
 #define CHASSIS_TOTAL_THERMAL_COUNT 8
 #define CHASSIS_FAN_COUNT     (CHASSIS_TOTAL_FAN_COUNT - CHASSIS_PSU_COUNT)
 #define CHASSIS_THERMAL_COUNT (CHASSIS_TOTAL_THERMAL_COUNT - CHASSIS_PSU_COUNT)
-
-#define PSU1_ID 1
-#define PSU2_ID 2
-
-#define PSU_MODULE_PREFIX "/bsp/module/psu%d_%s"
-#define PSU_POWER_PREFIX "/bsp/power/psu%d_%s"
-#define IDPROM_PATH "/bsp/eeprom/%s%d_info"
-
-#ifndef KERNEL_VERSION
-#define KERNEL_VERSION(a,b,c) (((a) << 16) + ((b) << 8) + (c))
-#endif
-
-/* LED related data
- */
-enum onlp_led_id
-{
-    LED_RESERVED = 0,
-    LED_SYSTEM,
-    LED_FAN1,
-    LED_FAN2,
-    LED_FAN3,
-    LED_FAN4,
-    LED_PSU,
-};
-
-typedef enum psu_type {
-    PSU_TYPE_UNKNOWN,
-    PSU_TYPE_AC_F2B,
-    PSU_TYPE_AC_B2F
-} psu_type_t;
-
-psu_type_t get_psu_type(int id, char* modelname, int modelname_len);
-
-int psu_read_eeprom(int psu_index, onlp_psu_info_t* psu_info,
-                     onlp_fan_info_t* fan_info);
-
-int onlp_fani_get_min_rpm(int id);
-int onlp_get_kernel_ver(void);
+#define CPLD_COUNT                 3
+#define SFP_PORT_COUNT            32
+#define CHASSIS_LED_COUNT       6
 
 #endif  /* __PLATFORM_LIB_H__ */
